Renaming react components
