+++ app/assets/stylesheets/provider/_header.scss
@@ -69,6 +69,9 @@
   .PopNavigation-trigger:hover i {
     color: $white;
   }
+  .PF4-avatar {
+    width: 30px;
+  }
 }
 .u-header-brand {
   align-items: center;
