+++ app/javascript/src/Navigation/components/header/Brand.jsx
@@ -1,7 +1,7 @@
 import React from 'react'
 import ContextSwitcher from 'Navigation/components/ContextSwitcherContainer'
 
-const SimpleBrand = () => <div className="u-header-brand">
+const Brand = () => <div className="u-header-brand">
   <div className="u-header-left-half">
     <div className='Header-logo'>Red Hat 3scale API Management</div>
   </div>
@@ -10,4 +10,4 @@ const SimpleBrand = () => <div className="u-header-brand">
   </div>
 </div>
 
-export default SimpleBrand
+export default Brand
